Split Converter into a converter and statement factory
To more cleanly separated the concerns, adjust the converter to only map
 ResultSet instances to required values for the JDBC DLQ. For
 statements, introduce a distinct "statement factory" interface and
 implementation. Chances are higher the statements need to be adjusted,
 for which the chance is slimmer for results.

#2350
